requirements.txt
#### Bot lib ####
git+https://gitlab.com/HarukaNetwork/OSS/PTB.git
telethon
#################
aiohttp>=2.2.5
bleach
emoji
future
googletrans>=3.1.0a0
hurry.filesize
markdown2
Pillow>=4.2.0
psycopg2-binary
pydantic
python-rapidjson
PyYAML
requests>=2.20.0
sqlalchemy
tldextract
wikipedia
covid
